+++ lib/noto_3d_emoji.dart
@@ -1,2 +1,4 @@
 export 'src/emoji_text.dart' show EmojiText;
 export 'src/noto_emoji.dart' show NotoEmoji;
+
+export 'src/coverage.dart' show notoCp, filterEmoji;